Add network tests for no connection
